[Variant] Caller provides ParentState to ValueBuilder methods (#8189)
# Which issue does this PR close?

- Closes https://github.com/apache/arrow-rs/issues/8188

# Rationale for this change

Today, `ValueBuilder::[try_]append_variant` unconditionally creates and
uses a `ParentState::Variant`, but that is incorrect when the caller is
a `ListBuilder` or `ObjectBuilder`. Rework the API so that the caller
passes their parent state, thus ensuring proper rollback in all
situations.

This is also a building block that will eventually let us simplify
`VariantArrayBuilder` to use a `ValueBuilder` directly, instead of a
`VariantBuilder`.

# What changes are included in this PR?

Several methods become associated functions.

# Are these changes tested?

Existing unit tests cover this refactor.

# Are there any user-facing changes?

No