[avro] Support all default types for avro schema's record field (#8210)
# Which issue does this PR close?

Closes https://github.com/apache/arrow-rs/issues/8209

# Rationale for this change

In the Field struct definition
```
/// A field within a [`Record`]
#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
pub struct Field<'a> {
    /// Name of the field within the record
    #[serde(borrow)]
    pub name: &'a str,
    /// Optional documentation for this field
    #[serde(borrow, default)]
    pub doc: Option<&'a str>,
    /// The field's type definition
    #[serde(borrow)]
    pub r#type: Schema<'a>,
    /// Optional default value for this field
    #[serde(borrow, default)]
    pub default: Option<&'a str>,
}
```
type is of type `Schema` whereas default is of type `str`. The default
should be supported for all types (e.g. int, array, map, nested record),
so we should make it more lenient.

More details on reproduction is mentioned in the Github Issue.

# What changes are included in this PR?

Relaxation of default type of avro scheam Field.

# Are these changes tested?

Added a unit test.

# Are there any user-facing changes?

It affects `pub struct Field` of `arrow-avro` package, but the impact
should be minimal as the `default` attribute is not being used.