[Variant] Allow appending raw object/list bytes to variant builders (#8141)
# Which issue does this PR close?

- Closes https://github.com/apache/arrow-rs/issues/8217

# Rationale for this change

When working with shredded variants, we need the ability to copy nested
object fields and array elements of one variant to a destination. This
is a cheap byte-wise copy that relies on the fact that the new variant
being built uses the same metadata dictionary as the source variant it
is derived from.

# What changes are included in this PR?

Define a helper macro that encapsulates the logic for variant appends,
now that we have three very similar methods (differing only in their
handling of list/object values and their return type).

Add new methods: `ValueBuilder::append_variant_bytes`, which is called
by new methods `VariantBuilder::append_value_bytes`,
`ListBuilder::append_value_bytes`, and
`ObjectBuilder::[try_]insert_bytes`.

# Are these changes tested?

New unit tests

# Are there any user-facing changes?

The new methods are public.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>