Add into_builder method for WriterProperties (#8272)
# Which issue does this PR close?

- Closes #8273 .

# Rationale for this change

When working with the library using encryption, we have sometimes found
it necessary to modify an existing set of `WriterProperties` on a
per-file basis to set specific encryption properties. More generally,
others may need to use an existing set of `WriterProperties` as a
template and modify the properties. I have implemented this feature by
adding an `into_builder` method, which appears to be the standard
approach in other parts of the library.


# Are these changes tested?

Yes, `test_writer_properties_builder` has been updated to add a
round-trip test for `into_builder`.

# Are there any user-facing changes?

Yes. `WriterProperties` now has a new `into_builder` method.

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>