feat: `SSLKEYLOGFILE` support for flight CLI (#8239)
# Which issue does this PR close?
\-

# Rationale for this change
This is #4875 now that the upstream changes are available.

Allows analysis of TLS traffic with an external tool like Wireshark.

See https://wiki.wireshark.org/TLS#using-the-pre-master-secret

# What changes are included in this PR?
New flag that opts into into the standard `SSLKEYLOGFILE` handling that
other libraries and browsers support.

# Are these changes tested?
Not automatic test, but I did validate that setting the flag AND the env
variable emits a log file that is successfully used by Wireshark to
decrypt the traffic.

# Are there any user-facing changes?
Mostly none for normal users, but might be helpful for developers.