[Variant] Remove boilerplate from make_shredding_row_builder (#8322)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #NNN.

# Rationale for this change

The type handling code is replicated twice in
`make_shredding_row_builder` -- for empty and non-empty paths,
respectively. The code is virtually identical in both cases, and
constructing any one row builder will always need two branches: (1)
check for empty path; and (2) dispatch on the correct type. It also uses
a macro to reduce boilerplate a bit.

# What changes are included in this PR?

Replace the macro with a new extension trait, and swap the control flow:
(1) dispatch on the correct type; and (2) check for empty path. This
cuts the affected code to 1/3 its original size.

# Are these changes tested?

Existing unit tests cover it.

# Are there any user-facing changes?

No.