Minor: avoid an `Arc::clone` in CacheOptions for Parquet PredicateCache (#8338)
# Which issue does this PR close?



# Rationale for this change

I found this while working on
https://github.com/apache/arrow-rs/pull/7997

Basically, the fact that the CacheOptionsBuilder had both owned and non
owned fields made it a bit akward to work with -- if it is going to be
zero copy (references) we pay the price of tracking lifetimes already.
We may as well just do so for the Arc as well

I don't expect this to make any measurable different in performance. I
am mostly treating it as a cleanup

# What changes are included in this PR?

Remove one Arc::clone

# Are these changes tested?

By existing CI
If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?

# Are there any user-facing changes?

If there are user-facing changes then we may require documentation to be
updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.