[Variant] ParentState tracks builder-specific state in a uniform way (#8324)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #NNN.

# Rationale for this change

The `ParentState` class, combined with `VariantBuilderExt` trait, makes
it pretty easy to work with variant builders. But it only works for
"well-known" builder types -- which does not and cannot include the
`VariantArrayBuilder` because it lives in a different crate.

This becomes a problem for e.g.
https://github.com/apache/arrow-rs/issues/8323, because it's currently
impossible to append multiple values to a `VariantArrayBuilder` -- it
needs to create and `finish` one`VariantArrayVariantBuilder` adapter for
each appended value.

Plus, we will eventually need a `VariantValueArrayBuilder` that works
with read-only metadata, for shredding, unshredding, and projecting
variant values. Which will undoubtedly encounter the same sorts of
problems, since shredding and unshredding code relies heavily on
`VariantBuilderExt`.

# What changes are included in this PR?

Make `ParentState` a customizable struct instead of an enum, with a
`BuilderSpecificState` that encapsulates the bits of finish and rollback
logic specific to each kind of builder. This allows
`VariantArrayBuilder` to directly implement `VariantBuilderExt`. It
simplifies both the array builder's implementation and the code that
uses it, and also opens the way for other custom builders like the
`VariantValueArrayBuilder` we will eventually need.

NOTE: One downside of this approach is the use of a boxed trait
instance. This effectively requires a heap allocation (and virtual
method dispatch) for every single value appended to a variant array,
which I don't love. However, none of our builder-using benchmarks show a
measurable slowdown.

If we don't like the overhead of the boxed trait approach, alternatives
we've considered include:
* Add new parent state enum variants for each new type of
`VariantBuilderExt`, even those that come from other crates.
   * PRO: The least amount of code of any alternative I've considered
   * PRO: Zero additional overhead compared to "native" types
* CON: Architectural violation to make parquet-variant crate (at least
somewhat) aware of parquet-variant-compute crate that depends on it.
* Make the various builder classes generic, and change `ParentState` to
a (not dyn-compat) trait that becomes a type constraint for those
classes.
   * NOTE: `VariantBuilderExt` is already not dyn-compat
* PRO: Even _less_ overhead than what we have today, because we no
longer need enum variant dispatch all over the place
* CON: A lot of code churn to make all the necessary classes generic.
Tho it's unclear how much that will actually impact users of the API.
Messy library code isn't necessarily bad, as long as it has a clean user
surface.
* Move the `VariantArrayBuilder` class into the `parquet-variant` crate
  * PRO: "fixes" the architectural violation
* CON: Gives `parquet-variant` a new `arrow-array` dependency
(currently, it only depends on `arrow-schema`).
* CON: Not flexible or future-proof -- anyone wishing to add a new kind
of builder must do it in the `parquet-variant` crate.

# Are these changes tested?

Yes, many unit tests were updated to use the new approach instead of the
old (removed) approach.

# Are there any user-facing changes?

No, because variant support is still experimental, but:
* `ParentState` becomes a struct that references a new public
`BuilderSpecificState` trait. All builders are updated to use it.
* `VariantArrayBuilder` now implements `VariantBuilderExt` directly, and
the old `VariantArrayVariantBuilder` adapter class has been removed.