[Variant] Rename VariantShreddingRowBuilder to VariantToArrowRowBuilder (#8344)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #NNN.

# Rationale for this change

The "shredding row builder" name is a bit hard to understand. And now
that we have `ArrowToVariantRowBuilder` (used by `cast_to_variant`), it
makes sense to rename these as `VariantToArrowRowBuilder` since that's
actually what they do.

While we're at it, rename `variant_get/mod.rs` as just `variant_get.rs`,
since it was the only file in its directory.

# What changes are included in this PR?

Just the rename (which includes a couple of file renames).

# Are these changes tested?

Pure code movement/rename -- compilation suffices

# Are there any user-facing changes?

No