[Variant] Add nullable arg to StructArrayBuilder::with_field (#8342)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #NNN.

# Rationale for this change

`StructArrayBuilder::with_field` is currently hard-wired to assume the
field will be nullable. This is unhelpful when adding non-nullable
fields such as variant a `metadata` column.

# What changes are included in this PR?

Add a third parameter, `nullable`. 

# Are these changes tested?

The builder is heavily used by unit tests, which have been adjusted to
pass the new param.

# Are there any user-facing changes?

No.

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>