[Variant] Make VariantToArrowRowBuilder an enum (#8345)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #NNN.

# Rationale for this change

Enum dispatch in rust is more efficient than virtual method dispatch,
and enums require far less boxing which makes them more memory efficient
as well. `ArrowToVariantRowBuilder` is already an enum, so it makes
sense for `VariantToArrowRowBuilder` to take the same approach.

# What changes are included in this PR?

Replace the trait with an enum.

# Are these changes tested?

Yes, existing row builder tests continue to pass.

# Are there any user-facing changes?

No.