[Variant] [Shredding] feat: Support typed_access for FixedSizeBinary (#8352)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8335

# Rationale for this change

See Issue

# What changes are included in this PR?

Support typed_access for FixedSizeBinary

# Are these changes tested?

Yes

# Are there any user-facing changes?

N/A. Variant support is still being developed