[Variant] [Shredding] Support typed_access for Utf8 and BinaryView (#8364)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8333

# Rationale for this change

See Issue

# What changes are included in this PR?

Support typed_access for Utf8 and BinaryView

# Are these changes tested?

Yes

# Are there any user-facing changes?

N/A. Variant support still in development

---------

Co-authored-by: Andrew Lamb <andrew@nerdnetworks.org>