[Variant] Implement new VariantValueArrayBuilder (#8360)
# Which issue does this PR close?

- Pre-work for https://github.com/apache/arrow-rs/issues/8361

# Rationale for this change

There is currently no good way to populate a new variant array with
variant values that reference an existing metadata, but that
functionality is needed when transforming existing variant data (e.g.
for shredding and unshredding operations).

# What changes are included in this PR?

Add a new `VariantValueArrayBuilder` that does not try to create new
metadata; instead, it wraps a `ReadOnlyMetadata` around the
`VariantMetadata` instance of the `Variant` value being inserted. This
takes advantage of the new generic `ParentState` capability.

NOTE: The new array builder does _not_ impl `VariantBuilderExt` because
it does not have a `MetadataBuilder` instance -- the instance is created
on demand as part of the insertion itself. Instead, callers can directly
invoke `VariantValueArrayBuilder::parent_state()`. This approach avoids
the considerable complexity of keeping an internal metadata column index
in sync with whatever external indexing might produce the variant value
to be appended. It also doesn't seem to matter -- I did some pathfinding
of variant shredding (going from binary to shredded variant based on
some target schema), and the `VariantBuilderExt` does not seem
especially helpful for that code.

# Are these changes tested?

New unit tests.

# Are there any user-facing changes?

New class.