Follow-up Improvements to Avro union handling  (#8385)
# Which issue does this PR close?

This work continues arrow-avro schema resolution support and aligns
behavior with the Avro spec.

- **Related to**: #4886 (“Add Avro Support”): ongoing work to round out
the reader/decoder, including schema resolution and type promotion.
- **Follow-ups/Context**: #8348 (Add arrow-avro Reader support for Dense
Union and Union resolution (Part 1))

# Rationale for this change

@scovich left a really solid
[review](https://github.com/apache/arrow-rs/pull/8348#pullrequestreview-3237862269)
on #8348 that wasn't completed until after the PR was merged in. This PR
is for addressing the suggestions and improving the code.

# What changes are included in this PR?

* Code quality improvements to `codec.rs`
* Improvements to `schema.rs` including spec compliant named type
errors.

# Are these changes tested?

1. No functionality was added / modified in `codec.rs` and all existing
tests are passing without changes.
2. Two new unit tests were added to `schema.rs` to cover the spec
compliant named type changes.

# Are there any user-facing changes?

N/A

---------

Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>