Add Arrow Variant Extension Type, remove  `Array` impl for `VariantArray` and `ShreddedVariantFieldArray` (#8392)
# Which issue does this PR close?

- closes https://github.com/apache/arrow-rs/issues/8319
- closes https://github.com/apache/arrow-rs/issues/8296

# Rationale for this change

This is needed to [read/write the Variant Parquet logical
type](https://github.com/apache/arrow-rs/issues/8370) and work with the
rest of the Arrow Ecosystem

Note, this is broken out the larger PR here: 
- from https://github.com/apache/arrow-rs/pull/8365


We need a way to write Variant encoded data to/from parquet, and the
current way the VariantArray is implemented doesn't work (panics when
writing to parquet). See tickets above

Instead of a `impl Array` it seems the better way to do this is using an
Arrow Extension Type. See
https://github.com/apache/arrow-rs/issues/8319#issuecomment-3285280543
for more details



# What changes are included in this PR?

1. remove the `Array` impl for `VariantArray`, which forces explict
conversions back/forth when reading/writing
2. remove the `Array` impl for `ShreddedVariantFieldArray`, which forces
explicit conversions back/forth when reading/writing
3. Add `VariantType` extension type

# Are these changes tested?

Yes, with new code and tests

# Are there any user-facing changes?

Yes, but this is not yet stable / released, so these changes have no
impact on the releasability of this code

---------

Co-authored-by: Ryan Johnson <scovich@users.noreply.github.com>
Co-authored-by: Matthijs Brobbel <m1brobbel@gmail.com>