Expose ReadPlan and ReadPlanBuilder (#8399)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8347 .

# Rationale for this change

Why are you proposing this change? If this is already explained clearly
in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand
your changes and offer better suggestions for fixes.

# What changes are included in this PR?

Expose relevant structs and methods for `ReadPlan` and `ReadPlanBuilder`
which can be used to build customized reader.

# Are these changes tested?

We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?

# Are there any user-facing changes?

Yes. Now `ReadPlanBuilder`, `ReadPlan` and relevant methods are exposed.

Signed-off-by: Ben Ye <benye@amazon.com>