[thrift-remodel] Use new Thrift encoder/decoder for Parquet page headers (#8376)
# Which issue does this PR close?
**Note: this targets a feature branch, not main**

- Part of #5854.

# Rationale for this change

This continues the remodel, moving on to `PageHeader` support.

# What changes are included in this PR?

Swaps out old `format` page header structs for new ones. This also adds
a `Read` based implementation of the thrift compact protocol reader (the
sizes of the Thrift encoded page headers are not knowable in advance, so
we need a way to read them from the thrift input stream used by the page
decoder).

This PR also makes decoding of the `Statistics` in the page header
optional (defaults to `false`). We do not use them, and the decode takes
a good chunk of time.

# Are these changes tested?

These changes should be covered by existing tests

# Are there any user-facing changes?

Yes, page level stats are no longer decoded by default