Update release instructions after `56.2.0` and `parquet_variant` tweaks (#8419)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Part of #7836 

# Rationale for this change

We added a new parquet_variant experimental feature which requires
publishing the crates. Thus add the necessary changes for publishing

# What changes are included in this PR?

1. Give parquet_variant a new version (it was published as version 0.1.0
-- see https://crates.io/crates/parquet-variant/0.1.0 etc)
2. Add instructions to publish the variant crates
3. fix readme path and versions in variant_compute_json

# Are these changes tested?
by CI and I tested them manually

# Are there any user-facing changes?

no