Quote `DataType::Struct` field names in `Display` formatting (#8291)
# Which issue does this PR close?
* Follows https://github.com/apache/arrow-rs/pull/8290 (merge that
first, and the diff of this PR will drop)
* https://github.com/apache/arrow-rs/pull/7469
* Part of https://github.com/apache/arrow-rs/issues/8351

# Rationale for this change
We would previously format structs like so:

`Struct(name1 type1, name2 nullable type2)`

This will break badly whenever the field name is anything but a simple
identifier. In other words: it allows [string
injection](https://xkcd.com/327/) if the field name contains an
end-paranthesis.

Except for that, it is also difficult to debug mistakingly bad field
names like " " or "\n".

# What changes are included in this PR?
We change the `Display` and `Debug` formatting of `Struct`

**Before**: `Struct(name1 type1, name2 nullable type2)`
**After**: `Struct("name1": type1, "name2": nullable type2)`

# Are these changes tested?
Yes - I've updated the existing tests.

# Are there any user-facing changes?
Yes, changing the `Display` formatting is a **breaking change**