Unpin comfytable (#8440)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes https://github.com/apache/arrow-rs/pull/8318
- follow on to https://github.com/apache/arrow-rs/pull/8244

# Rationale for this change

Now that we have updated the MSRV we can unpin the comfy table version
- https://github.com/apache/arrow-rs/pull/8429

# What changes are included in this PR?


# Are these changes tested?



If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?

# Are there any user-facing changes?

If there are user-facing changes then we may require documentation to be
updated before approving the PR.

If there are any breaking changes to public APIs, please call them out.

Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>