[Variant]: Implement `DataType::FixedSizeList` support for `cast_to_variant` kernel (#8282)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8281.

# Rationale for this change

# What changes are included in this PR?

Support the last DataType in `cast_to_variant` kernel

# Are these changes tested?

Yes

# Are there any user-facing changes?

New type supported