refactor: split `num` dependency (#8459)
# Which issue does this PR close?
\-

# Rationale for this change
`num` is a meta-crate that bundles functionality of a bunch of `num-*`
crates, similar to how `futures` work:

```text
num v0.4.3
├── num-bigint v0.4.6
│   ├── num-integer v0.1.46
│   │   └── num-traits v0.2.19 (*)
│   └── num-traits v0.2.19 (*)
├── num-complex v0.4.6
│   └── num-traits v0.2.19 (*)
├── num-integer v0.1.46 (*)
├── num-iter v0.1.45
│   ├── num-integer v0.1.46 (*)
│   └── num-traits v0.2.19 (*)
│   [build-dependencies]
│   └── autocfg v1.5.0
├── num-rational v0.4.2
│   ├── num-bigint v0.4.6 (*)
│   ├── num-integer v0.1.46 (*)
│   └── num-traits v0.2.19 (*)
└── num-traits v0.2.19 (*)
```

We don't need all these sub-crates but only a very specific set. So
instead of using the meta-crate, let's use the actual things we need.

# What changes are included in this PR?
Dependency changes.

# Are these changes tested?
It still compiles.

# Are there any user-facing changes?
Faster compilation.