Add Parquet geospatial statistics utility (#8414)
# Which issue does this PR close?

We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax.

- Closes #8411.

# Rationale for this change

The presence of relevant statistics when writing geometries and/or
geographies is one of the primary motivations behind the GEOMETRY and/or
GEOGRAPHY in Parquet. We'd like to make it easy for writers to provide
them!

# What changes are included in this PR?

This PR introduces `Interval` and `WraparoundInterval` structs that
handle interval math, and a `GeometryBounder` that iterates over input
using the fantastic `wkb` crate (via `geo-traits`).

# Are these changes tested?

Yes!

# Are there any user-facing changes?

All public structures and functions are documented (although I am not
sure what the final public API will be).

---------

Co-authored-by: Kyle Barron <kylebarron2@gmail.com>