[Variant] Add a VariantBuilderExt impl for VariantValueArrayBuilder (#8501)
# Which issue does this PR close?

- Relates to https://github.com/apache/arrow-rs/issues/8336

# Rationale for this change

`VariantValueArrayBuilder` was already a bit inconvenient to work with,
because it provides no implementation of `VariantBuilderExt`.
Pathfinding for variant unshredding exposed this as an actual problem
and not merely an inconvenience.

# What changes are included in this PR?

Add a thin wrapper for `VariantValueArrayBuilder` for which we can `impl
VariantBuilderExt` -- thus allowing the value builder to participate
fully in the variant builder ecosystem.

# Are these changes tested?

Yes, existing unit tests updated to use the new capability.

# Are there any user-facing changes?

New pub method `VariantValueArrayBuilder::as_builder_ext`