[Variant] ReadOnlyMetadataBuilder borrows its underlying VariantMetadata (#8502)
# Which issue does this PR close?

- Relates to https://github.com/apache/arrow-rs/issues/8336

# Rationale for this change

While pathfinding variant unshredding code, I noticed that there's a lot
of variant metadata cloning going on. It's a 32-byte object, so cloning
is not free. And it turns out that ~all read-only metadata builder usage
in the code base can work just as easily with a borrowed reference
because ultimately we're working with bytes borrowed from a binary view
that holds the metadata column.

# What changes are included in this PR?

Update the `ReadOnlyMetadataBuilder` to borrow the `VariantMetadata` it
uses, and update call sites accordingly.

# Are these changes tested?

Yes, the change affects several unit tests.

# Are there any user-facing changes?

Signature of `ReadOnlyMetadataBuilder::new` has changed.