[Variant] Introduce new BorrowedShreddingState concept (#8499)
# Which issue does this PR close?

- Relates to https://github.com/apache/arrow-rs/issues/8336

# Rationale for this change

While pathfinding support for unshredding variant objects and arrays, I
ran into the same issue that `variant_get` already suffers --
`ShreddingState` is inconvenient because it owns the `value` and
`typed_value` columns it refers to, even tho borrowing them usually
suffices.

# What changes are included in this PR?

Define a new `BorrowedShreddingState` which does what it says, and
update `ShreddedPathStep` (in variant_get.rs) to use it.

Also, make the constructor fallible in order to correctly report casting
failures if the `value` column is not a binary view.

# Are these changes tested?

Yes, existing tests cover this code.

# Are there any user-facing changes?

New `TryFrom` and `From` implementations.

`impl From<&StructArray> for ShreddingState` was replaced by a suitable
`TryFrom`.