[thrift-remodel] Remove most usage of `parquet::format` structures (#8476)
# Which issue does this PR close?
**Note: this targets a feature branch, not main**

- Part of #5854.

# Rationale for this change

This should complete the major work of the remodel. Follow-on tasks
include performance tweaks, documentation, and adding the ability to
skip unneeded structures when decoding. None of the latter should
involve breaking changes.

# What changes are included in this PR?

The major change is removing all of the code to convert to
`parquet::format` structures. The only places those are still used are
in the benchmark and `bin` code which are not strictly in the parquet
crate. Once those are cleaned up we can deprecate the `format` module.

This also adds a markdown file documenting the use of the new Thrift
macros.

# Are these changes tested?

Should be covered by existing tests.

# Are there any user-facing changes?

Yes

---------

Co-authored-by: Matthijs Brobbel <m1brobbel@gmail.com>