Remove allow unused from arrow-avro lib.rs file (#8493)
# Which issue does this PR close?

- **Related to**: #4886 (“Add Avro Support”)

**NOTE:** This PR is stacked on
https://github.com/apache/arrow-rs/pull/8492

# Rationale for this change

`arrow-avro` has seen significant development since `#![allow(unused)]`
was temporarily added to `lib.rs`. Due to fast iteration on the code,
this has led to unused methods and imports throughout the crate, which
need to be cleaned up prior to `arrow-avro` becoming public.

This PR simply removes `#![allow(unused)]` and cleans up the
`arrow-avro` crate's code to comply.

# What changes are included in this PR?

Deleted the `#![allow(unused)]` in `lib.rs` and updated the crate's code
as needed. This impacted almost every files of the crate, however the
changes in this PR are 100% focused and isolated around only the work
related to removing `#![allow(unused)]`.

# Are these changes tested?

The changes in this PR are covered by existing tests. No new
functionality or behavior has been changed/added. This PR is simply
clean up around removing `#![allow(unused)]` from `lib.rs`.

# Are there any user-facing changes?

N/A