Arrow-avro Writer Dense Union support  (#8550)
# Which issue does this PR close?

Relates to:
https://github.com/apache/arrow-rs/pull/8348
https://github.com/apache/arrow-rs/issues/4886

# Rationale for this change

This PR completes the efforts of @jecsand838, adding dense union support
to the encoder side of the crate, along with four other minor extensions
of existing time-related encoding.

Note: currently this PR is stacked behind
https://github.com/apache/arrow-rs/pull/8546. Once that's merged this
will be updated and will not include those changes.

# What changes are included in this PR?

- Dense union support for the writer
- Tests

# Are these changes tested?

- A full round-trip test, reading in an existing union avro file and
asserting that the output matches expectations
- Unit tests covering new encoders.

# Are there any user-facing changes?

Crate not yet public

---------

Co-authored-by: Connor Sanders <connor@elastiflow.com>
Co-authored-by: Connor Sanders <170039284+jecsand838@users.noreply.github.com>
Co-authored-by: Matthijs Brobbel <m1brobbel@gmail.com>