Support writing GeospatialStatistics in Parquet writer (#8524)
# Which issue does this PR close?

- Closes #8523.

# Rationale for this change

One of the primary reasons the GeoParquet community was excited about
first-class Parquet Geometry/Geography support was the built-in column
chunk statistics (we had a workaround that involved adding a struct
column, but it was difficult for non-spatial readers to use it and very
difficult for non-spatial writers to write it). This PR ensures it is
possible for arrow-rs to write files that include those statistics.

# What changes are included in this PR?

This PR inserts the minimum required change to enable this support.

# Are these changes tested?

Yes!

# Are there any user-facing changes?

There are several new functions (which include documentation).
Previously it was difficult or impossible to actually write Geometry or
Geography logical types, and so it is unlikely any previous usage would
be affected.

---------

Co-authored-by: Ed Seidl <etseidl@users.noreply.github.com>